\name{yeah-package}
\alias{yeah-package}
\alias{yeah}
\docType{package}
\title{yeah}
\description{yeah package connects R users with voice.}
\tabular{ll}{
Package: \tab yeah\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-03-02\cr
License: \tab MIT\cr
}
}
\author{dichika}
\examples{
yeah("yeah")
}
